pyproject.toml
[tool.poetry]
name = "bulk_image_rename"
version = "0.1.0"
description = ""
authors = ["Thorne Wolfenbarger <thornecanyon@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
openai = "^1.5.0"
google-cloud-aiplatform = "^1.38.1"
google-generativeai = "^0.3.1"
pillow = "^10.1.0"
retry = "^0.9.2"
click = "^8.1.7"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
